requirements.txt
# Data handling
pandas>=2.1

# Environment variables
python-dotenv>=1.0

# LLM
openai>=1.54.0

# YOLO11 Training Dependencies
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Data processing
numpy>=1.26
scikit-learn>=1.3.0
PyYAML>=6.0